requirements1.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7
python-dotenv==1.0.1
sympy==1.13.0
numpy==1.26.4
